html/canvas/element/path-objects/2d.path.ellipse.angle.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.path.ellipse.angle</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.ellipse.angle</h1>
<p class="desc">StartAngle and EndAngle should be based on the eccentric circle</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("StartAngle and EndAngle should be based on the eccentric circle");
_addTest(function(canvas, ctx) {

ctx.ellipse(50, 50, 50, 25, Math.PI*0.3, 0, Math.PI*1.2, false);
ctx.fill();

_assertPixel(canvas, 46,18, 0,0,0,0);
_assertPixel(canvas, 32,13, 0,0,0,255);


});
</script>
